requirements.txt
pandas
numpy
scipy
python-dateutil